+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/pay_button/PayButtonViewModel.kt
@@ -11,6 +11,7 @@ import com.mercadopago.android.px.internal.callbacks.Event
 import com.mercadopago.android.px.internal.callbacks.PaymentServiceEventHandler
 import com.mercadopago.android.px.internal.core.ConnectionHelper
 import com.mercadopago.android.px.internal.core.ProductIdProvider
+import com.mercadopago.android.px.internal.extensions.isNotNullNorEmpty
 import com.mercadopago.android.px.internal.features.explode.ExplodeDecoratorMapper
 import com.mercadopago.android.px.internal.features.pay_button.PayButton.OnReadyForPaymentCallback
 import com.mercadopago.android.px.internal.features.pay_button.UIProgress.*
@@ -83,6 +84,9 @@ internal class PayButtonViewModel(
         if (connectionHelper.checkConnection()) {
             handler?.prePayment(object : OnReadyForPaymentCallback {
                 override fun call(paymentConfiguration: PaymentConfiguration, confirmTrackerData: ConfirmData?) {
+                    if(paymentConfiguration.customOptionId.isNotNullNorEmpty()) {
+                        paymentSettingRepository.clearToken()
+                    }
                     startSecuredPayment(paymentConfiguration, confirmTrackerData)
                 }
             })